Rename API events to more appropriate names. Call load event before data is set (to allow modification), call save event but ignore inventory from event. This essentially makes the event just a cancellable flag but it atleast lets a dev get the final inventory being saved (I would have just cloned and sent it and passed the inventory as an argument for the event but Bukkit strikes again).
BackpackHandler will need to be observed when the GUI is constructed to
see the impact on the FileSystem when saving. Either the GUI needs to load
in-memory backpacks or load in-memory backpacks and call a save to the
FileSystem when the GUI closes (should be able to prevent exploits better
that way as well).

Signed-off-by: Chris Sanders <NinjaZidane@hotmail.com>